requirements.txt
pygame
numpy
cv2
pillow
python-dotenv
